# gate/src/main/java/ru/mironov/gate/controller/StudentGateController.java
package ru.mironov.gate.controller;

import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;
import ru.mironov.gate.api.StudentGateApi;
import ru.mironov.gate.model.StudentGateCreateRequest;
import ru.mironov.gate.model.StudentGateResponse;
import ru.mironov.gate.client.api.StudentDataApi;
import ru.mironov.gate.client.model.StudentDataCreateRequest;
import ru.mironov.gate.client.model.StudentDataResponse;

@RestController
@RequiredArgsConstructor
public class StudentGateController implements StudentGateApi {

    private final StudentDataApi studentsFeignClient;

    @Override
    public ResponseEntity<StudentGateResponse> createStudent(StudentGateCreateRequest request) {
        StudentDataCreateRequest dataRequest = new StudentDataCreateRequest();
        dataRequest.setFullName(request.getFullName());
        dataRequest.setPassport(request.getPassport());

        StudentDataResponse dataResponse = studentsFeignClient.createStudentDataInData(dataRequest);

        StudentGateResponse gateResponse = new StudentGateResponse();
        gateResponse.setId(dataResponse.getId());
        gateResponse.setFullName(dataResponse.getFullName());
        gateResponse.setPassport(dataResponse.getPassport());

        return ResponseEntity.status(201).body(gateResponse);
    }

    @Override
    public ResponseEntity<StudentGateResponse> getStudent(Long id) {
        try {
            StudentDataResponse dataResponse = studentsFeignClient.getStudentDataByIdFromData(id);

            StudentGateResponse gateResponse = new StudentGateResponse();
            gateResponse.setId(dataResponse.getId());
            gateResponse.setFullName(dataResponse.getFullName());
            gateResponse.setPassport(dataResponse.getPassport());

            return ResponseEntity.status(200).body(gateResponse);
        } catch (Exception e) {
            return ResponseEntity.notFound().build();
        }
    }
}